Create main.yml
The Panda is one of the globes most adorable creatures and look even more adorable when chilling with a tasty piece of bamboo. The Panda is creature/agent of peace and has fostered enduring friendship between nations. It's native land is China but it has found home in various nations on almost all the world's continent. A creature as adorable as it is peaceful. Always minding its own business and chilling with bamboo. It must be preserved .